+++ sites/geohub/src/components/pages/map/layers/raster/RasterLayer.svelte
@@ -11,15 +11,13 @@
 		CLASSIFICATION_METHOD_CONTEXT_KEY,
 		COLORMAP_NAME_CONTEXT_KEY,
 		LAYERLISTSTORE_CONTEXT_KEY,
-		LEGEND_READONLY_CONTEXT_KEY,
 		NUMBER_OF_CLASSES_CONTEXT_KEY,
 		RASTERRESCALE_CONTEXT_KEY,
 		createClassificationMethodStore,
 		createColorMapNameStore,
 		createNumberOfClassesStore,
 		createRasterRescaleStore,
-		type LayerListStore,
-		type LegendReadonlyStore
+		type LayerListStore
 	} from '$stores';
 	import { createEventDispatcher, getContext, setContext } from 'svelte';
 
@@ -29,7 +27,6 @@
 	export let isExpanded: boolean;
 
 	const layerListStore: LayerListStore = getContext(LAYERLISTSTORE_CONTEXT_KEY);
-	const legendReadonly: LegendReadonlyStore = getContext(LEGEND_READONLY_CONTEXT_KEY);
 
 	const rescaleStore = createRasterRescaleStore();
 	setContext(RASTERRESCALE_CONTEXT_KEY, rescaleStore);
@@ -81,9 +78,7 @@
 </script>
 
 <LayerTemplate {layer} bind:isExpanded on:toggled={handleToggleChanged}>
-	{#if !$legendReadonly}
-		<Tabs bind:tabs bind:activeTab on:tabChange={(e) => (activeTab = e.detail)} />
-	{/if}
+	<Tabs bind:tabs bind:activeTab on:tabChange={(e) => (activeTab = e.detail)} />
 
 	<div class="panel-content px-2 pb-2">
 		<div hidden={activeTab !== TabNames.LEGEND}>
@@ -93,13 +88,10 @@
 				bind:tags={layer.dataset.properties.tags}
 			/>
 		</div>
-		{#if !$legendReadonly && !isRgbTile}
+		{#if !isRgbTile}
 			<div hidden={activeTab !== TabNames.TRANSFORM}>
 				<RasterTransform bind:layer />
 			</div>
 		{/if}
 	</div>
 </LayerTemplate>
-
-<style lang="scss">
-</style>